add test link to assertion
