+++ src/api/network_health_check.rs
@@ -7,7 +7,7 @@ use crate::{graphql::QueryNetworkId, MinaMesh};
 impl MinaMesh {
   pub async fn network_health_check(self, network_identifier: NetworkIdentifier) -> Result<bool> {
     let QueryNetworkId { network_id } = self.graphql_client.send(QueryNetworkId::build(())).await?;
-    if network_identifier.blockchain == "MINA" {
+    if network_identifier.blockchain == "mina" {
       unimplemented!();
     }
     if network_identifier.network == network_id {
