--- tests/compare/block.rs
@@ -1,16 +1,0 @@
-use mina_mesh::models::{BlockRequest, NetworkIdentifier, PartialBlockIdentifier};
-
-use super::CompareGroup;
-
-pub fn block<'a>() -> CompareGroup<'a> {
-  ("/block", vec![
-    Box::new(BlockRequest {
-      network_identifier: Box::new(NetworkIdentifier::new("mina".to_string(), "debug".to_string())),
-      block_identifier: Box::new(PartialBlockIdentifier::new()),
-    }),
-    Box::new(BlockRequest {
-      network_identifier: Box::new(NetworkIdentifier::new("mina".to_string(), "debug".to_string())),
-      block_identifier: Box::new(PartialBlockIdentifier { index: Some(52676), hash: None }),
-    }),
-  ])
-}
